api/requirements.txt
asgiref==3.8.1
blinker==1.9.0
click==8.1.7
Flask==3.1.0
graphql-core==3.2.5
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
six==1.16.0
SQLAlchemy==2.0.36
strawberry-graphql==0.251.0
types-PyYAML==6.0.12.20240917
typing_extensions==4.12.2
Werkzeug==3.1.3
